+++ src/scene.c
@@ -68,7 +68,7 @@ void FreeScene(Scene *const scene)
     FreePlayer(&scene->player);
 }
 
-const int *GetSlicesPlayerScene(const Scene *const scene)
+const float *GetSlicesPlayerScene(const Scene *const scene)
 {
     return GetSlicesPlayer(&scene->player);
 }
